refactor(agent-query): clarify session infrastructure (#565)
